PREAPPS-502: Add unit tests for `getFrequency`.
